move localhost tests
