+++ integration/tests/localhost/localhost-different-port-different-instance.test.ts
@@ -14,7 +14,7 @@ import { expect, test } from '@playwright/test';
 import type { Application } from '../../models/application';
 import type { FakeUser } from '../../testUtils';
 import { createTestUtils } from '../../testUtils';
-import { prepareApplication } from './utils';
+import { prepareApplication } from '../sessions/utils';
 
 test.describe('multiple apps running on localhost using different Clerk instances @sessions', () => {
   test.describe.configure({ mode: 'serial' });
